+++ autokeras/hypermodel/base.py
@@ -246,107 +246,3 @@ class Head(Block):
     def postprocess(self, y):
         """Postprocess the output of the Keras Model."""
         return y
-
-
-class HyperBlock(Block):
-    """HyperBlock uses hyperparameters to decide inner Block graph.
-
-    A HyperBlock should be build into connected Blocks instead of individual Keras
-    layers. The main purpose of creating the HyperBlock class is for the ease of
-    parsing the graph for preprocessors. The graph would be hard to parse if a Block,
-    whose inner structure is decided by hyperparameters dynamically, contains both
-    preprocessors and Keras layers.
-
-    When the preprocessing layers of Keras are ready to cover all the preprocessors
-    in AutoKeras, the preprocessors should be handled by the Keras Model. The
-    HyperBlock class should be removed. The subclasses should extend Block class
-    directly and the build function should build connected Keras layers instead of
-    Blocks.
-
-    # Arguments
-        output_shape: Tuple of int(s). Defaults to None. If None, the output shape
-            will be inferred from the AutoModel.
-        name: String. The name of the block. If unspecified, it will be set
-            automatically with the class name.
-    """
-
-    def __init__(self, output_shape=None, **kwargs):
-        super().__init__(**kwargs)
-        self.output_shape = output_shape
-
-    def build(self, hp, inputs=None):
-        """Build the HyperModel instead of Keras Model.
-
-        # Arguments
-            hp: HyperParameters. The hyperparameters for building the model.
-            inputs: A list of instances of Node.
-
-        # Returns
-            An Node instance, the output node of the output Block.
-        """
-        raise NotImplementedError
-
-
-class Preprocessor(Block):
-    """Hyper preprocessing block base class.
-
-    It extends Block which extends Hypermodel. A preprocessor is a Hypermodel, which
-    means it is a search space. However, different from other Hypermodels, it is
-    also a model which can be fit.
-    """
-
-    def build(self, hp):
-        """Get the values of the required HyperParameters.
-
-        It does not build and return a Keras Model, but initialize the
-        HyperParameters for the preprocessor to be fit.
-        """
-        pass
-
-    def update(self, x, y=None):
-        """Incrementally fit the preprocessor with a single training instance.
-
-        # Arguments
-            x: EagerTensor. A single instance in the training dataset.
-            y: EagerTensor. The targets of the tasks. Defaults to None.
-        """
-        raise NotImplementedError
-
-    def transform(self, x, fit=False):
-        """Incrementally fit the preprocessor with a single training instance.
-
-        # Arguments
-            x: EagerTensor. A single instance in the training dataset.
-            fit: Boolean. Whether it is in fit mode.
-
-        Returns:
-            A transformed instanced which can be converted to a tf.Tensor.
-        """
-        raise NotImplementedError
-
-    def output_types(self):
-        """The output types of the transformed data, e.g. tf.int64.
-
-        The output types are required by tf.py_function, which is used for transform
-        the dataset into a new one with a map function.
-
-        # Returns
-            A tuple of data types.
-        """
-        raise NotImplementedError
-
-    @property
-    def output_shape(self):
-        """The output shape of the transformed data.
-
-        The output shape is needed to build the Keras Model from the AutoModel.
-        The output shape of the preprocessor is the input shape of the Keras Model.
-
-        # Returns
-            A tuple of int(s) or a TensorShape.
-        """
-        raise NotImplementedError
-
-    def finalize(self):
-        """Training process of the preprocessor after update with all instances."""
-        pass
